+++ lib/lightstreamer/stream_connection.rb
@@ -118,6 +118,7 @@ module Lightstreamer
 
       buffer = StreamBuffer.new
       options[:response_block] = -> (data, _, _) { buffer.process data, &method(:process_stream_line) }
+      options[:expects] = 200
 
       Excon.post url, options
     rescue Excon::Error => error